Removed temporary fix for lowercase code gen
The PR (https://github.com/kubernetes/kubernetes/pull/68484) was merged
into upstream kubernetes.  This commit removes the temporary fix for
lowercasing.  The latest version of the code gen has an update to the
patch client that use a parameter that is not in current version of
Go-Client used by this repo so dropped code gen for patching. Patching
is not being used currently so safe to drop.  May consider dropping
other verbs not used in the client in future.